refactor: flatten account panel structure
