+++ src/components/SecuritySection.tsx
