+++ src/components/modals/TopUpModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { X, Plus } from 'lucide-react'
 
-import type { BillingData, PaymentMethod } from '../../../types/account'
+import type { BillingData, PaymentMethod } from '../../types/account'
 
 interface TopUpModalProps {
   showTopUpModal: boolean
