+++ src/hooks/useAccountData.ts
@@ -1,10 +1,12 @@
-import { useExists, useTypedFile } from '@artifact/client/hooks'
+import { useExists, useJson } from '@artifact/client/hooks'
 import { accountDataSchema } from '../types/account'
+import type { AccountData } from '../types/account'
 import { useMemo } from 'react'
 
 const useAccountData = () => {
   const exists = useExists('profile.json')
-  const data = useTypedFile('profile.json', accountDataSchema)
+  const raw = useJson('profile.json') as unknown
+  const data = raw ? (accountDataSchema.parse(raw) as AccountData) : undefined
 
   const loading = useMemo(
     () => exists === null || (exists && data === undefined),
