--- src/views/index.ts
@@ -1,1 +1,0 @@
-export { default as AccountView } from './AccountView'
